+++ mypy.ini
@@ -10,5 +10,5 @@ files =
     dapr/clients/**/*.py,
     dapr/conf/**/*.py,
     dapr/serializers/**/*.py,
-    flask_dapr/actor/*.py,
+    flask_dapr/actor.py,
     examples/demo_actor/**/*.py
